Fix signedness warnings in contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c
contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c: In function ‘gdk_pixbuf_xlib_get_from_drawable’:
contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c:1268:38: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
  g_return_val_if_fail (src_x + width <= src_width
                                      ^~
/usr/include/glib-2.0/glib/gmacros.h:427:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c:1268:2: note: in expansion of macro ‘g_return_val_if_fail’
  g_return_val_if_fail (src_x + width <= src_width
  ^~~~~~~~~~~~~~~~~~~~
contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c:1269:28: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
          && src_y + height <= src_height, NULL);
                            ^~
/usr/include/glib-2.0/glib/gmacros.h:427:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c:1268:2: note: in expansion of macro ‘g_return_val_if_fail’
  g_return_val_if_fail (src_x + width <= src_width
  ^~~~~~~~~~~~~~~~~~~~

+++ contrib/gdk-pixbuf-xlib/gdk-pixbuf-xlib-drawable.c
@@ -1265,8 +1265,8 @@ gdk_pixbuf_xlib_get_from_drawable (GdkPixbuf *dest,
 	}
 
 	g_return_val_if_fail (src_x >= 0 && src_y >= 0, NULL);
-	g_return_val_if_fail (src_x + width <= src_width
-			      && src_y + height <= src_height, NULL);
+	g_return_val_if_fail ((guint) (src_x + width) <= src_width &&
+			      (guint) (src_y + height) <= src_height, NULL);
 
 	if (dest) {
 		g_return_val_if_fail (dest_x >= 0 && dest_y >= 0, NULL);
